tests: split introspect and runtime
