+++ src/docker/docker-compose.yml
@@ -1,4 +1,4 @@
-version: '3.7'
+version: "3.7"
 services:
   # In Docker
   # - TEST_POSTGRES_URI=postgres://prisma:prisma@postgres/
@@ -18,7 +18,7 @@ services:
     volumes:
       - postgres:/var/lib/postgresql/data
     ports:
-      - '5432:5432'
+      - "5432:5432"
 
   mysql:
     image: mysql:8.0
@@ -32,7 +32,7 @@ services:
     volumes:
       - mysql:/var/lib/mysql
     ports:
-      - '3306:3306'
+      - "3306:3306"
 
   mariadb:
     image: mariadb:10
@@ -45,7 +45,7 @@ services:
     volumes:
       - mariadb:/var/lib/mysql
     ports:
-      - '4306:3306'
+      - "4306:3306"
 
 volumes:
   postgres:
