+++ src/packages/cli/package.json
@@ -70,6 +70,7 @@
     "@prisma/sdk": "workspace:*",
     "@prisma/studio": "0.288.0",
     "@prisma/studio-server": "0.288.0",
+    "@sindresorhus/slugify": "^1.1.0",
     "@types/debug": "4.1.5",
     "@types/jest": "26.0.13",
     "@types/pg": "7.14.4",
@@ -122,7 +123,7 @@
     "test:sqlite": "jest integration.sqlite --forceExit",
     "test:postgresql": "jest integration.postgresql --forceExit",
     "test:mysql": "jest integration.mysql --forceExit",
-    "test:mariadb": "jest integration.mariadb --forceExit",
+    "test:mariadb": "jest --forceExit integration.mariadb",
     "test:commands": "./fixtures/test.sh && jest dotenv- format doctor generate studio && jest introspect --forceExit",
     "test": "pnpm run test:commands && pnpm run test:integration",
     "test:integration": "jest integration --forceExit",
