+++ src/packages/cli/src/__tests__/integration.postgresql.test.ts
@@ -1,10 +1,10 @@
 import * as PG from 'pg'
-import { integrationTest } from './__helpers__/integrationTest'
+import { runtime } from './__helpers__/integrationTest'
 
 const connectionString =
   process.env.TEST_POSTGRES_URI || 'postgres://prisma:prisma@localhost:5432/'
 
-integrationTest<PG.Client>({
+runtime<PG.Client>({
   database: {
     name: 'postgresql',
     datasource: {
