+++ src/packages/cli/src/__tests__/integration/mariadb/runtime.test.ts
@@ -1,0 +1,8 @@
+import * as IntegrationTest from '../../__helpers__/integrationTest'
+import database from './database'
+import scenarios from './scenarios'
+
+IntegrationTest.runtime({
+  database,
+  scenarios,
+})
